Empty all tables with delete for trilogy
Following on from https://github.com/rails/rails/pull/56097, move
the `empty_all_tables` implementation into the abstract MySQL adapter
so that it is also used with the `trilogy` adapter.
