+++ activerecord/test/cases/adapters/mysql2/mysql2_adapter_test.rb
@@ -352,25 +352,3 @@ class Mysql2AdapterTest < ActiveRecord::Mysql2TestCase
       super(@conn, "ex", definition, &block)
     end
 end
-
-class Mysql2AdapterEmptyAllTablesTest < ActiveRecord::Mysql2TestCase
-  self.use_transactional_tests = false
-
-  def setup
-    @conn = ActiveRecord::Base.lease_connection
-  end
-
-  def test_empty_all_tables_uses_delete_not_truncate
-    queries = []
-    subscriber = ActiveSupport::Notifications.subscribe("sql.active_record") do |event|
-      queries << event.payload[:sql] if event.payload[:name] == "Delete Tables"
-    end
-
-    @conn.empty_all_tables
-
-    ActiveSupport::Notifications.unsubscribe(subscriber)
-
-    assert queries.any? { |q| q.include?("DELETE FROM") }, "Expected DELETE statements to be used"
-    assert queries.none? { |q| q.include?("TRUNCATE") }, "Expected TRUNCATE statements to NOT be used"
-  end
-end
